refactor(#zimic): abstract worker access resources utilities
