+++ packages/zimic/src/interceptor/http/interceptorWorker/__tests__/HttpinterceptorWorker.node.test.ts
@@ -1,15 +1,12 @@
-import { describe, expect } from 'vitest';
+import { describe } from 'vitest';
 
 import Server from '@/cli/server/Server';
-import { getCrypto } from '@/utils/crypto';
-import { joinURL } from '@/utils/fetch';
+import { getNodeAccessResources } from '@tests/utils/workers';
 
 import { declareSharedHttpInterceptorWorkerTests } from './shared/workerTests';
 
-describe('HttpInterceptorWorker (Node.js)', async () => {
+describe('HttpInterceptorWorker (Node.js)', () => {
   const server = new Server();
-
-  const crypto = await getCrypto();
 
   declareSharedHttpInterceptorWorkerTests({
     platform: 'node',
@@ -18,24 +15,8 @@ describe('HttpInterceptorWorker (Node.js)', async () => {
       await server.start();
     },
 
-    getBaseURL(type) {
-      if (type === 'local') {
-        return {
-          baseURL: 'http://localhost:3000',
-          pathPrefix: '',
-        };
-      }
-
-      const hostname = server.hostname();
-      const port = server.port()!;
-      expect(port).not.toBe(null);
-
-      const pathPrefix = `path-${crypto.randomUUID()}`;
-
-      return {
-        baseURL: joinURL(`http://${hostname}:${port}`, pathPrefix),
-        pathPrefix,
-      };
+    getAccessResources(type) {
+      return getNodeAccessResources(type, server);
     },
 
     async stopServer() {
